Separate DBContext Cliente configuration
That is an other aproach to implements the configuration to dbcontext
from EFCore.
In this implementation is possible to segregate all configuration in a
class with an unique responsability, just for configure a context.
After We have a class with all configuration We need to informe to our
DbContex to apply the config class.